+++ site/lib/src/layouts/dash_layout.dart
@@ -2,15 +2,18 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:convert';
+
 import 'package:jaspr/jaspr.dart';
 import 'package:jaspr_content/jaspr_content.dart';
 
 import '../client/global_scripts.dart';
-// import '../components/cookie_notice.dart';
-import '../components/footer.dart';
-import '../components/header.dart';
-import '../components/sidenav.dart';
+import '../components/common/client/cookie_notice.dart';
+import '../components/layout/footer.dart';
+import '../components/layout/header.dart';
+import '../components/layout/sidenav.dart';
 import '../models/sidenav_model.dart';
+import '../style_hash.dart';
 import '../util.dart';
 
 /// The base Jaspr Content layout for wrapping site content.
@@ -19,6 +22,8 @@ abstract class DashLayout extends PageLayoutBase {
 
   @override
   String get name;
+
+  List<String> get defaultBodyClasses => [];
 
   @override
   @mustCallSuper
@@ -194,7 +199,17 @@ abstract class DashLayout extends PageLayoutBase {
 
     return Component.fragment(
       [
-        if (bodyClass != null) Document.body(attributes: {'class': bodyClass}),
+        const Document.html(
+          attributes: {
+            'lang': 'zh-cmn-Hans',
+            'dir': 'ltr',
+          },
+        ),
+        Document.body(
+          attributes: {
+            'class': [?bodyClass, ...defaultBodyClasses].toClasses,
+          },
+        ),
         //         raw('''
         // <noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-5VSZM5J" height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
         // '''),